Upisano jos korisnih podatka u skriptu i male izmene u klasama modela. #67a

+++ src/main/java/com/tim18/bolnicar/dto/AppointmentDTO.java
@@ -10,7 +10,7 @@ import java.util.Date;
 public class AppointmentDTO {
     private Integer id;
     private Date datetime;
-    private Integer duration;
+    private Double duration;
     private Double discount;
     private Room room;
     private ExaminationType type;
@@ -21,7 +21,7 @@ public class AppointmentDTO {
 
     public AppointmentDTO(Integer id,
                           Date datetime,
-                          Integer duration,
+                          Double duration,
                           Double discount,
                           Room room,
                           ExaminationType type,
@@ -71,11 +71,11 @@ public class AppointmentDTO {
         this.datetime = datetime;
     }
 
-    public Integer getDuration() {
+    public Double getDuration() {
         return duration;
     }
 
-    public void setDuration(Integer duration) {
+    public void setDuration(Double duration) {
         this.duration = duration;
     }
